Rename module to secp256k1hdwallet
