+++ packages/launchpad/src/lcdapi/gov.spec.ts
@@ -4,7 +4,7 @@ import { sleep } from "@cosmjs/utils";
 import { coins } from "../coins";
 import { assertIsBroadcastTxSuccess } from "../cosmosclient";
 import { makeSignDoc } from "../encoding";
-import { Secp256k1Wallet } from "../secp256k1wallet";
+import { Secp256k1Wallet } from "../secp256k1hdwallet";
 import { SigningCosmosClient } from "../signingcosmosclient";
 import {
   dateTimeStampMatcher,
